Remove unuseful question from the beginning
When the skil starts it asked you to start the test and then asked you again.
