+++ custom/intents/personality/index.ts
@@ -1,5 +1,3 @@
 export * from './personalityTest'
-export * from './startTest'
-export * from './denyTest'
 export * from './testInProgress'
 export * from './checkMyPersonality'
