+++ custom/locales/translations.ts
@@ -2,9 +2,9 @@ export const locales = {
     'en-US': {
         translation: {
             "SKILL_NAME": "Hello world",
-            "WELCOME_MSG": "Welcome to the Probot, you can say 'test' to discover your personality!",
+            "WELCOME_MSG": "Welcome to the Probot. Probot helps you to discover your personality by starting the test in any moment.",
             "GOODBYE_MSG": "Goodbye!",
-            "HELLO_MSG": "Hello world!",
+            "HELLO_MSG": "If you want to start the test, simply ask `start the test`. In any moment yout can check your personality.",
             "HELP_MSG": "You can say hello to me!",
             "ERROR_MSG": "Sorry, I can't understand the command. Please say again.",
             "ERROR_UNEXPECTED_MSG": "Sorry, an unexpected error has occured. Please try again later.",
